requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
google-api-core==2.23.0
google-auth==2.36.0
google-cloud-vision==3.8.1
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.16.0
urllib3==2.2.3
